chore: FeatureC refactoring
