+++ packages/scoped-elements/test-web/components/lit/index.js
@@ -2,7 +2,6 @@ import '../vanilla/globally-defined-button.js';
 
 export { FeatureA } from './FeatureA.js';
 export { FeatureB } from './FeatureB.js';
-export { FeatureC } from './FeatureC.js';
 export { ScopeOnInstanceLevel } from './ScopeOnInstanceLevel.js';
 export { WithoutStaticScopedElements } from './WithoutStaticScopedElements.js';
 export { AdoptStyles } from './AdoptStyles.js';
